fix footprint and bbx always appears together issue (#70)
* fix footprint and bbx always appears together issue

* CHANGELOG

---------
