+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+- **PODAAC-5857**
+  - Fixed the issue so footprint and bbx do not always appear together while isoXmlSpatial is configured to footprint only
+  - remove request.close() statement in CMRLambdaRestClient so the response could be pulled out and logged properly instead of always logging apache chunk read error due to the http channel was closed.
+  - adding more parameters to Jsonschema2pojo plugin to generate proper pojo from ummg schema
 ### Security
 
 ## [8.5.0]
