+++ src/main/java/gov/nasa/cumulus/metadata/aggregator/IsoGranule.java
@@ -20,7 +20,7 @@ public class IsoGranule extends UMMGranule {
     private List<String> inputGranules;
     private String PGEVersionClass;
 
-    private IsoType isoType;
+
 
     public IsoGranule() {
         this.identifiers = new HashMap<>();
@@ -129,13 +129,5 @@ public class IsoGranule extends UMMGranule {
 
     public void setPGEVersionClass(String PGEVersionClass) {
         this.PGEVersionClass = PGEVersionClass;
-    }
-
-    public IsoType getIsoType() {
-        return isoType;
-    }
-
-    public void setIsoType(IsoType isoType) {
-        this.isoType = isoType;
     }
 }
